cahnged to entitynot found, and a lot

+++ src/main/java/com/villysiu/yumtea/service/impl/MilkServiceImpl.java
@@ -3,11 +3,13 @@ package com.villysiu.yumtea.service.impl;
 import com.villysiu.yumtea.models.tea.Milk;
 import com.villysiu.yumtea.repo.tea.MilkRepo;
 import com.villysiu.yumtea.service.MilkService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Map;
+import java.util.NoSuchElementException;
 
 @Service
 @RequiredArgsConstructor
@@ -16,8 +18,8 @@ public class MilkServiceImpl implements MilkService {
     private final MilkRepo milkRepo;
 
     @Override
-    public Milk updateMilk(Long id, Map<String, Object> milkDto) throws RuntimeException {
-        Milk milk = milkRepo.findById(id).orElseThrow(()->new RuntimeException("Milk not found."));
+    public Milk updateMilk(Long id, Map<String, Object> milkDto) {
+        Milk milk = milkRepo.findById(id).orElseThrow(()->new EntityNotFoundException("Milk not found."));
         for(Map.Entry<String, Object> milkDtoEntry : milkDto.entrySet()) {
             String field = milkDtoEntry.getKey();
             Object value = milkDtoEntry.getValue();
@@ -38,9 +40,9 @@ public class MilkServiceImpl implements MilkService {
         return milk;
     }
 
-    public Milk getMilkById(Long id) throws RuntimeException {
+    public Milk getMilkById(Long id) {
         return milkRepo.findById(id)
-                .orElseThrow(()->new RuntimeException("Milk not found."));
+                .orElseThrow(()->new EntityNotFoundException("Milk not found."));
     }
 
     @Override
@@ -49,14 +51,15 @@ public class MilkServiceImpl implements MilkService {
     }
 
     @Override
-    public Milk createMilk(Milk milk) throws RuntimeException {
+    public Milk createMilk(Milk milk){
         return milkRepo.save(milk);
     }
 
     @Override
-    public String deleteMilk(Long id) throws RuntimeException {
-        milkRepo.findById(id)
-                .orElseThrow(()->new RuntimeException("Milk not found."));
+    public String deleteMilk(Long id) {
+        if (!milkRepo.existsById(id)) {
+            throw new EntityNotFoundException("Milk not found");
+        }
         milkRepo.deleteById(id);
         return "Milk deleted";
     }
